Add files via upload
Just throwing an idea out there as a starting point for how we structure our video. I suggest we introduce a user story (10 seconds max) as a lead into a 1-2 minute demonstration of each of the features we want to highlight. These slides were more of an excuse to play around with Dalle 2 than a serious suggestion to use this 'art'.